Address usability issues on mobile
PDS app bar is wider than some mobile device screens and clickable
elements are too close together.

Hardcoded placement of information text caused it to extend past the
screen (and had it misaligned with the information icon). Determine
placement by the width of the title/logo, which can change dependeing on
the device.

Enlarge the information icon and extend the space between it and the
adjacent drop-down menu.

See #9.