change dropdown link to dropdown button
- remove focus capability from dropdown container
- remove unnecessary span element from button
- replace unwieldy caret (plain styling instead of CSS code)
- make entire button clickable, not just the caret

https://pds-atmospheres.nmsu.edu/data_and_services/atmospheres_data/catalog.htm
and like pages have javascript which colors certain links red. Rather
than also resorting to javascript, and because the behavior better fits
a button rather than a link, I changed the dropdown link to a button.

For accessibility, a button must be focusable, so I removed the focus
from the container and applied the style and behavior to the new button.
The way the caret is presented (formerly with a CSS code) had to be
changed because its shape extended beyond the button, changing the shape
of the outline when focused. The span element was removed because it
seemed unnecessary. If I had done it before for a specific reason, I
don't recall it, and this'll (hopefully) teach me to document better.

Resolves #31